Fix bug: All parameters after a default parameter must have a default, too
Also: Reorder googletests so that tests concerning function declarations are grouped together
